fix: stringify to hex
Calling `toString()` on a buffer might lead to the same result
although the buffers are different. Example:

    $ node
    > const buf1 = Buffer.from('15603533e990dfde', 'hex')
    undefined
    > const buf2 = Buffer.from('15603533e990dfe0', 'hex')
    undefined
    > buf1.toString()
    '\u0015`53���'
    > buf2.toString()
    '\u0015`53���'
    > buf1.toString() === buf2.toString()
    true
    > buf1.toString('hex') === buf2.toString('hex')
    false

Currently CIDs are stored internally with their `toString()` representation.
This could fail as different CIDs could end up with the same representation.
The fix is to use the hex representation instead.

Closes #186.

BREAKING CHANGE: emitted keys change

Bitswap has an event emitter which emits events when a block is received
and when a block is not wanted anymore. These events contain a string
representation of the CID. This representation changes with this commit.

Though practically it shouldn't have any impact on any system, except if
you have hard-coded specific CIDs.