fix: use consistent encoding for cid comparison
BREAKING CHANGE: Emitted events have different bytes

The emitted events contain the stringified version of the CID, as we
change it to the base encoding the CID has, those bytes may be different
to previous versions of this module.

Though this shouldn't have any impact on any other modules as the
events are only used internally.