fix: blocks may not be in the blockstore yet
Because some blockstore implementations batch up write before committing
them to disk or wherever, they may not be available when we think
they should be.

I noticed this module throwing unhandledPromiseRejections around this
so the change here handles missing blocks by putting the task back
into the queue to process later, with a fairly arbitrary 5x limit on
processing a given task.