fix: re-sort queue after adding tasks
When tasks are added to an existing list of tasks for a given peer,
we need to sort the queue to ensure the order is correct, otherwise
we never process pending tasks as task lists with pending tasks need
to be moved up the queue.

Fixes the build problems exposed in ipfs/js-ipfs#2992

Also upgrade aegir to a safe version.