fix: do not rebroadcast want list
We should not need to rebroadcast our want lists, because we send
it to newly connected peers.

Rebroadcasting is a throwback to when the network was less reliable,
but if the connection is unreliable it is dropped and a new connection
established, which triggers sending the whole wantlist as the remote
is seen as a newly connected peer.

fixes #160