fix: pass peer id to onPeerConnect
The persistent peer store in libp2p 0.28 means sometimes we have peers
when we start up, so we try to start bitswap for them when we start up
if we have an open connection already.

The connection object has a remotePeer property that should be used
instead of passing the whole thing to `this._onPeerConnect`.